Final новая функциональность

+++ src/utils/Api.js
@@ -32,8 +32,8 @@ export class Api {
             method: 'PATCH',
             headers: this.headers,
             body: JSON.stringify({
-                name: newUserInfo.nameProfile,
-                about: newUserInfo.textProfile
+                name: newUserInfo.name,
+                about: newUserInfo.about
             })
         })
     }
@@ -67,28 +67,29 @@ export class Api {
         })
     }
 
-    //метод добавления лайка у карточки
-    addLike(id) {
-        return this._sendRequest(`/cards/likes/${id}`, {
-            method: 'PUT',
-            headers: this.headers
-        })
-    }
-
-    //метод удаления лайка
-    deleteLike(id) {
-        return this._sendRequest(`/cards/likes/${id}`, {
-            method: 'DELETE',
-            headers: this.headers
-        })
+    //метод добавляет или удаляет лайк
+    changeLikeCardStatus(id, isLiked) {
+        if (isLiked) {
+            return this._sendRequest(`/cards/likes/${id}`, {
+                method: 'PUT',
+                headers: this.headers
+            })
+        } else {
+            return this._sendRequest(`/cards/likes/${id}`, {
+                method: 'DELETE',
+                headers: this.headers
+            })
+        }
     }
 
     //метод удаления карточки
-    deleteCard(id) {
+    deleteCard(id, isOwn) {
+        if (!isOwn) {
         return this._sendRequest(`/cards/${id}`, {
             method: 'DELETE',
             headers: this.headers
         })
+      }
     }
 }
